temporal bases and mpGLM initiation
the temporal basis functions are initiated more robustly. If the
specification is "none," the number of bases is fixed to be one.

The initiation of the mixture of Poisson generalized linear model is
changed such that the weights of all filters of all neurons in the first state are
negative ones and in the second state ones. Previously, the weights are
randomly initiated between -1 and 0 in the first state and 0 and 1 in
the second state. This reduces the variability in the initialization,
but makes sure that the norm weights of no neuron in either state is too
close to 0, which can cause the weights to all collapse to zero.