more restricted range of prior precisions
previously, the range of the precision prior was ∈ [0.05, 1e6]. This
resulted in a Hessian of the posterior that had eigenvalues that ranged
diversely in magnitudes, which meant that the Hessian was poorly
conditioned (the condition number of a matrix is the ratio of the
absolute values of the largest and smallest eigenvalue). Because
optimization of the posterior was performed using L-BFGS, whose
efficiency depended on well-conditioned Hessian matrix, it is critical
that the Hessian's eigenspectrum is more limited.

In addition, previously there was a bug in the code for optimizing the
evidence. The parameters whose prior's precision was not being learned
was not being updated! That means the baseline firing rate of each
neuron was not being updated! This is fixed.

An additional modification of the evidence optimization procedure may
help it to converge. Previously, in each outer iteration of the evidence
procedure, the parameters of the model were initialized as the
approximate MAP values. Now, the parameters are the actual MAP values.