filter options
For the Poisson mixture GLM of each neuron, the post-spike,
post-stereoclick, and pre-movement filters can be excluded by setting
"tbf_hz" to "NaN". Moreover,
the gaussian prior on the weight of the temporal basis functions of
these filters can be assigned separate ranges.

post-modification tests are documented in the repository "tzluo", in
"tzluo/analyses/analyses_2022_10_17a_test"