new options for evidence optimization
New options were introduced to specify which group of hyperparameters
are learned and which are not. These groups include the precision on the
drift-diffusion parameters, the precision on the weight of the
post-stereoclick filter, the precision on the weight of the pre-movement
filter, the precision on the weight of the post-photostimulus filter,
the precision on the weight of the post-spike filter, the precision on
the weight of the accumulator (𝐯), and the precision on the
state-dependent change in the weight of the accumulator (Δ𝐯).

Tests were performed by 'analysis_2022_12_18b_test'.