separation of gain and post-stereoclick filter
The parameters specifying the gain of each neuron's Poisson mixture
generalized linear model and the model's time-varying response to the
stereoclick are now separate, in order that the temporal richness of the
response to the stereoclick may be regularized without concomittant
regularization of the baseline gain of the model.

The functions implementing the concatentation of model parameters from a
composite to a vector and the sorting of concatenated values back into a
composite have been modularized.

Tests were performed by "analysis_2022_12_30a_test".