corrected sampling from Poisson mixture GLM
When the filters of the Poisson mixture generalized linear models were
redefined to separate the gain from the post-stereoclick kernel, an
error had been introduced in the sampling. The error consists of
duplicating the input from the post-spike filter and results in
predictions of peri-stimulus time histograms that do not match the
observations. The error has been corrected.