revert to using 𝛃 from using Δ𝐯
The pre- and postcommitment weights could be parametrized in at least
two separate ways. In both parametrizations, the pre-commitment weight
is represented by the parameter 𝐯. For the post-commitment weight, it
could be parametrized with a separate parameter 𝛃, or as a change from
the pre-commitment weight, as 𝐯 + Δ𝐯. Initially, state-dependent
encoding was parametrized with (𝐯, 𝛃), and I changed the parametrization
to be (𝐯, 𝐯 + Δ𝐯). I am reverting to the initial parametrization because
I suspect that the older parametrization is better for model fitting.