choiceLL_scaling_exponent
tests were documented in `analysis_2023_02_01a_test`

I wish to identify an approach for computing the scale factor of the
log-likelihood such that not only the psychometric function can be well
captured, but also the spike trains are leveraged to learn the height of
the absorbing bound of the drift-diffusion process. Moreover, I wish the
scale factor to be not so large that improvements in fitting the spike
trains can be directly observed in improvements in fitting the
psychometric function. Currently, the log-likelihood of the choices are
multiplied by a scale factor $s$ equal to the trial average of the
product of the number of neurons $N$ by the number of time steps $T$ in
a trial.

$$s \equiv \frac{1}{M}\sum_{m=1}^M N_mT_m$$

This scale factor allows the psychometric function to be well captured.
However, this is scale factor is so large that the spike trains do not
contribute to learning the drift-diffusion parameters.

Therefore, I modified the code base so that a fixed hyperparameter, $a$,
is introduced in the computation of the scale factor. The scale factor
is equal to the trial average above raised to the exponent $a$:

$$s_{new} \equiv \left(\frac{1}{M}\sum_{m=1}^M N_mT_m\right)^a$$

Note that when $a=1$, the new scale factor is equal to the old scale
factor, and when $a=0$, no scaling occurs.