initial
The goal of this branch is to clean up the package to retain only the
functionalities that contributed to the manuscript of Luo, Kim et al.
(2023). Moreover, the observation model is to be modified to be more
similar to the observation model in FINDR.

Functionalities to remove
- coupling

Changes to the observation model
- the drift, i.e., the across-trial component of the bias, will be
  estimated using firing rates before the trial rather than using time
in session
- the relative contributions of the within- and across-trial kernels to
  the bias will be estimated separately from the parameters controlling
the latent dynamics.